pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chronix"
version = "0.0.0"
description = "Chronix Discord bot (development)"
dependencies = [
  "discord.py>=2.0.0",
  "aiohttp>=3.8.0",
  "pydantic>=1.10.0",
  "python-dotenv>=0.21.0",
  "asyncpg>=0.27.0",
  "pytest>=7.0.0",
  "pytest-asyncio>=0.22.0",
]
